Code clean up

+++ src/controller/profileController.ts
@@ -4,7 +4,10 @@ import { User, UserToDo } from '../../models';
 import fs = require('fs');
 
 export const updateProfile = async (req, res, next) => {
+    console.log("WWW");
+
     try {
+        console.log("WWW");
         const { firstName, lastName, email, deleteImage } = req.body
         const USER_ID = req.payLoad.aud
         const user = await User.findOne({ where: { id: parseInt(USER_ID) } })
@@ -20,7 +23,18 @@ export const updateProfile = async (req, res, next) => {
         if (email) {
             user.email = email
         }
-        if (deleteImage == "true" && user.dataValues.profileimage) {
+        //* Upload Image Logic
+        let imagePath = req?.file?.path || null
+        if (user?.dataValues?.profileimage && imagePath) {
+            if (fs.existsSync(user?.dataValues?.profileimage)) {
+                fs.unlinkSync(user.dataValues.profileimage);
+            }
+        }
+        if (imagePath) {
+            user.profileimage = imagePath
+        }
+
+        if (deleteImage == "true" && user.dataValues.profileimage && !imagePath) {
             if (fs.existsSync(user?.dataValues?.profileimage)) {
                 fs.unlinkSync(user.dataValues.profileimage);
                 user.profileimage = null
@@ -29,33 +43,8 @@ export const updateProfile = async (req, res, next) => {
         await user.save()
         res.send({ user })
     } catch (error) {
-        console.log("createAccount error", error);
-        if (error.isJoi === true) error.status = 422
-        next(error)
-    }
-}
+        console.log("error", error);
 
-export const updateImage = async (req, res, next) => {
-    try {
-        const USER_ID = req.payLoad.aud
-        let imagePath = req?.file?.path || null
-        const user = await User.findOne({ where: { id: parseInt(USER_ID) } })
-        if (user?.dataValues?.profileimage) {
-            if (fs.existsSync(user?.dataValues?.profileimage)) {
-                fs.unlinkSync(user.dataValues.profileimage);
-            }
-        }
-        if (!user) {
-            throw createError.Conflict("No User Found")
-        }
-        if (imagePath) {
-            user.profileimage = imagePath
-        }
-        await user.save()
-        res.send({ user })
-    } catch (error) {
-        console.log("login error", error);
-        if (error.isJoi === true) error.status = 422
         next(error)
     }
 }
@@ -71,10 +60,8 @@ export const deleteProfile = async (req, res, next) => {
         if (!todo) {
             throw createError.NotExtended("No Data with us")
         }
-        console.log("All Todo", JSON.stringify(todo));
         const ogData = JSON.stringify(todo)
         const jsonData = JSON.parse(ogData)
-        console.log("jsonData", jsonData);
         for (let i = 0; i < jsonData.length; i++) {
             let currentItem = jsonData[i]
             if (currentItem?.todoImage) {
@@ -96,7 +83,6 @@ export const deleteProfile = async (req, res, next) => {
         }
         res.send({ deleted: true });
     } catch (error) {
-        console.log("login error", error);
         if (error.isJoi === true) error.status = 422
         next(error)
     }
